[profiler] for async RPC, use separate api in RecordFunction to handle nested scopes.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33030

We previously had an issue (described in #32517, #32884) where async RPCs would not be profiled correctly if there were nested scopes during profiling (nested scopes can be created by with torch.autograd.profiler.record_function("foo")...)

The cause of this was that the logic for handling nested scopes was not correct when we invoked RecordFunction::end() from a separate thread.

As proposed in #32884, this change adds 2 APIs to RecordFunction:
runEndCallbacks() which just runs the end callbacks for profiling, and doesn't reset thread local RecordFunction data
resetThreadLocalFunc that resets the thread local RecordFunction data.

For profiling RPCs, we call resetThreadLocalFunc in the creating thread to avoid the above issue.
tested by modifying unit test to add the nested scope.
ghstack-source-id: 97817370

Differential Revision: [D19739387](https://our.internmc.facebook.com/intern/diff/D19739387/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D19739387/)!